api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2023.3
pandas==2.1.1